Added more TaskService Tests
