+++ src/test/java/de/thowl/tnt/core/TestTaskService.java
@@ -17,7 +17,12 @@
  */
 package de.thowl.tnt.core;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.util.List;
+import java.util.Date;
 
 import org.junit.jupiter.api.Test;
 
@@ -27,12 +32,15 @@ import org.springframework.test.context.ActiveProfiles;
 
 import de.thowl.tnt.core.services.TaskService;
 import de.thowl.tnt.storage.TaskRepository;
+import de.thowl.tnt.storage.UserRepository;
+import de.thowl.tnt.storage.entities.User;
+import de.thowl.tnt.storage.entities.Task;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @SpringBootTest
 @ActiveProfiles("test")
-public class TestTaskServiceInt {
+public class TestTaskService {
 
 	@Autowired
 	private TaskService tasksvc;
@@ -40,13 +48,62 @@ public class TestTaskServiceInt {
 	@Autowired
 	private TaskRepository tasks;
 
+	@Autowired
+	private UserRepository users;
+
 	@Test
-	void testSetDone() {
-		log.info("entering test testSetDone");
+	void testToggleDone() {
+		log.info("entering test testToggleDone");
 
 		this.tasksvc.toggleDone(1);
 
 		assertTrue(this.tasks.findById(1).isDone(), "This should be true");
+
+		this.tasksvc.toggleDone(1);
+
+		assertFalse(this.tasks.findById(1).isDone(), "This should be false");
+	}
+
+	@Test
+	public void testAddTask() {
+
+		log.info("entering test testAddTask");
+
+		long taskAmount;
+
+		taskAmount = this.tasks.count();
+
+		this.tasksvc.add("Tasktester", "Test Task", "Task Content", "High", new Date(), new Date());
+
+		assertEquals(taskAmount + 1, this.tasks.count());
+	}
+
+	@Test
+	public void testDeleteTask() {
+
+		log.info("entering test testAddTask");
+
+		long taskAmount;
+
+		taskAmount = this.tasks.count();
+
+		this.tasksvc.delete(1);
+
+		assertEquals(taskAmount - 1, this.tasks.count());
+	}
+
+	@Test
+	public void testGetAllTasks() {
+
+		User user;
+		List<Task> tasks;
+		List<Task> result;
+
+		user = this.users.findByUsername("Tasktester");
+		tasks = this.tasks.findByUser(user);
+		result = this.tasksvc.getAllTasks("Tasktester");
+
+		assertEquals(tasks.size(), result.size(), "These should be the same size");
 	}
 
 }
